+++ package.json
@@ -6,7 +6,7 @@
   "bin": "bin.js",
   "scripts": {
     "precommit": "lint-staged",
-    "start": "node-dev example",
+    "start": "node --stack_trace_limit=100 -r trace -r clarify example",
     "style:lint": "standard",
     "style:format": "prettier-standard '**/*.js' '**/*.md'",
     "style": "npm-run-all -s style:format style:lint",
@@ -46,14 +46,15 @@
   "homepage": "https://github.com/buttcloud/longshore#readme",
   "devDependencies": {
     "ava": "^0.25.0",
+    "clarify": "^2.1.0",
     "dependency-check": "^2.7.0",
     "husky": "^0.14.3",
     "lint-staged": "^7.0.5",
-    "node-dev": "^3.1.3",
     "npm-run-all": "^4.0.1",
     "nyc": "^10.1.2",
     "prettier-standard": "^8.0.1",
-    "standard": "^11.0.1"
+    "standard": "^11.0.1",
+    "trace": "^3.1.0"
   },
   "dependencies": {
     "ansi-escape-sequences": "^4.0.0",
