+++ specs/util/namespace.js
